requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Document Processing
python-docx>=1.1.0
pypdf>=3.17.0
PyPDF2

# LangChain and AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-google-genai>=0.0.6

# PDF Generation
reportlab>=4.0.0


# Data visualization
pandas>=1.5.0
plotly>=5.15.0


